+++ algorithm_distillation/__init__.py
@@ -1,3 +1,4 @@
-from algorithm_distillation import rl_tasks
+from algorithm_distillation import models, tasks
+from algorithm_distillation.ad import AlgorithmDistillation, GymAD
 
-__all__ = ["rl_tasks"]
+__all__ = ["models", "tasks", "AlgorithmDistillation", "GymAD"]
