--- algorithm_distillation/rl_tasks/__init__.py
@@ -1,5 +1,0 @@
-from algorithm_distillation.rl_tasks.ad import AlgorithmDistillation, GymAD
-from algorithm_distillation.rl_tasks.task import GymTask, Task
-from algorithm_distillation.rl_tasks.task_manager import TaskManager
-
-__all__ = ["AlgorithmDistillation", "GymAD", "Task", "GymTask", "TaskManager"]
