+++ algorithm_distillation/tasks/rl/__init__.py
@@ -1,0 +1,4 @@
+from algorithm_distillation.tasks.rl.task import GymTask, Task
+from algorithm_distillation.tasks.rl.task_manager import TaskManager
+
+__all__ = ["Task", "GymTask", "TaskManager"]
